net: move testing utilities in separate module
