+++ src/devices/src/virtio/net/event_handler.rs
@@ -115,8 +115,10 @@ impl Subscriber for Net {
 
 #[cfg(test)]
 pub mod tests {
-    use super::*;
-    use crate::virtio::net::device::tests::{NetEvent, NetQueue, TestHelper};
+    use crate::check_metric_after_block;
+    use crate::virtio::net::test_utils::test::TestHelper;
+    use crate::virtio::net::test_utils::{NetEvent, NetQueue};
+    use logger::{Metric, METRICS};
 
     #[test]
     fn test_event_handler() {
