+++ src/devices/src/virtio/net/mod.rs
@@ -12,11 +12,11 @@ pub const RX_INDEX: usize = 0;
 // The index of the tx queue from Net device queues/queues_evts vector.
 pub const TX_INDEX: usize = 1;
 
-#[macro_use]
 pub mod device;
 pub mod event_handler;
 pub mod persist;
 mod tap;
+pub mod test_utils;
 
 pub use self::device::Net;
 pub use self::event_handler::*;
